backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
alembic==1.13.2
bcrypt==4.2.0
python-dotenv==1.0.1
pinecone-client==5.0.1
google-generativeai==0.8.3
pydantic==2.9.2
requests==2.32.3
uvicorn==0.32.0
Werkzeug==3.0.4
supabase==2.6.0
pypdf==4.3.1
gunicorn==23.0.0
sentence-transformers==2.7.0
torch==2.7.1
